+++ docs/changelog.rst
@@ -11,6 +11,7 @@ v1.2.0 - 2016-xx-xx
 - Improve data export field names (`#132 <https://github.com/dennissiemensma/dsmr-reader/issues/132>`_).
 - Display average temperature in archive (`#122 <https://github.com/dennissiemensma/dsmr-reader/issues/122>`_).
 - Pie charts on trends page overlap their canvas (`#136 <https://github.com/dennissiemensma/dsmr-reader/issues/136>`_).
+- 'Slumber' consumption (`#115 <https://github.com/dennissiemensma/dsmr-reader/issues/115>`_).
 
 
 v1.1.2 - 2016-05-01